moved ssda_sync.py and updated imports
